WIP: Migrate changes to the main hardware integration
